backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
redis==5.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
APScheduler==3.10.4
Werkzeug==3.0.1
eventlet==0.33.3
